+++ src/main/java/com/zerobase/oriticket/domain/members/entity/UserEntity.java
@@ -1,5 +1,6 @@
 package com.zerobase.oriticket.domain.members.entity;
 
+import com.zerobase.oriticket.domain.members.constants.MemberStatus;
 import com.zerobase.oriticket.domain.members.constants.RoleType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,7 +9,10 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Temporal;
+import jakarta.persistence.TemporalType;
 import java.sql.Timestamp;
+import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -20,30 +24,43 @@ import org.hibernate.annotations.CreationTimestamp;
 @AllArgsConstructor
 @Builder
 @Entity
-public class Admin {
+public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long adminId;
+    private long id;
 
-    @Column(nullable = false, length = 30, unique = true)
+    @Column(length = 30, unique = true)
     private String email;
 
-    @Column(nullable = false, length = 30)
+    @Column(length = 30)
     private String name;
 
-    @Column(nullable = false, length = 30, unique = true)
+    @Column(length = 30, unique = true)
     private String nickname;
 
-    @Column(nullable = false, length = 100)
+    @Column( length = 100)
     private String password;
 
-    @Column(length = 10)
-    @Enumerated(EnumType.STRING)
-    private RoleType roles = RoleType.ADMIN;
+    @Column()
+    @Temporal(TemporalType.TIMESTAMP)
+    private LocalDateTime birthDate;
 
-    @Column(nullable = false)
+    @Column(length = 30, unique = true)
+    private String phoneNum;
+
+    @Enumerated(EnumType.STRING)
+    private RoleType role = RoleType.ROLE_USER;
+
+    @Enumerated(EnumType.STRING)
+    private MemberStatus status = MemberStatus.ACTIVE;
+
     @CreationTimestamp
     private Timestamp registeredAt;
+
+    @CreationTimestamp
+    private Timestamp modifiedAt;
+
+    private String oauth; // kakao,naver
 
 }
