+++ src/main/java/com/zerobase/oriticket/domain/members/security/JwtAuthenticationFilter.java
@@ -1,0 +1,2 @@
+package com.zerobase.oriticket.domain.members.security;public class JwtAuthenticationFilter {
+}
