+++ src/main/java/com/zerobase/oriticket/domain/members/security/SecurityConfiguration.java
@@ -1,0 +1,2 @@
+package com.zerobase.oriticket.domain.members.security;public class SecutiryConfiguration {
+}
