--- src/main/java/com/zerobase/oriticket/global/dto/ErrorResponse.java
@@ -1,10 +1,0 @@
-package com.zerobase.oriticket.global.dto;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class ErrorResponse {
-    private String message;
-}
