--- src/main/java/com/zerobase/oriticket/global/exception/AbstractException.java
@@ -1,6 +1,0 @@
-package com.zerobase.oriticket.global.exception;
-
-public abstract class AbstractException extends RuntimeException{
-    abstract public int getErrorCode();
-    abstract public String getMessage();
-}
