--- src/main/java/com/zerobase/oriticket/global/exception/CustomExceptionHandler.java
@@ -1,20 +1,0 @@
-package com.zerobase.oriticket.global.exception;
-
-import com.zerobase.oriticket.global.dto.ErrorResponse;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-
-@ControllerAdvice
-public class CustomExceptionHandler {
-
-    @ExceptionHandler(AbstractException.class)
-    protected ResponseEntity<?> handleCustomException(AbstractException e) {
-        ErrorResponse errorResponse = ErrorResponse.builder()
-                .message(e.getMessage())
-                .build();
-
-        return new ResponseEntity<>(errorResponse, HttpStatus.resolve(e.getErrorCode()));
-    }
-}
